requirements.txt
uvicorn
fastapi
pydantic
astropy
starlette
pyserial
